+++ podcasts/EpisodeFilterOverlayController.swift
@@ -34,12 +34,15 @@ class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableVi
         setCurrentStatus()
 
         navigationController?.navigationBar.setValue(true, forKey: "hidesShadow")
-        addCloseButton()
 
         if FeatureFlag.playlistsRebranding.enabled {
+            navigationItem.largeTitleDisplayMode = .always
+
             handleThemeChanged()
 
             saveButton.setTitle(L10n.playlistSmartRuleSaveButton, for: .normal)
+        } else {
+            addCloseButton()
         }
     }
 
@@ -159,6 +162,14 @@ class EpisodeFilterOverlayController: FilterSettingsOverlayController, UITableVi
         if FeatureFlag.playlistsRebranding.enabled {
             saveButton.backgroundColor = AppTheme.colorForStyle(.primaryInteractive01)
             changeNavTint(titleColor: nil, iconsColor: AppTheme.colorForStyle(.primaryIcon03))
+        }
+    }
+
+    override func dismissViewController() {
+        if FeatureFlag.playlistsRebranding.enabled {
+            navigationController?.popViewController(animated: true)
+        } else {
+            dismiss(animated: true, completion: nil)
         }
     }
 }
