+++ podcasts/PodcastFilterOverlayController.swift
@@ -70,14 +70,17 @@ class PodcastFilterOverlayController: PodcastChooserViewController, PodcastSelec
     }
 
     func setupNavBar() {
-        setupCloseButton()
         if FeatureFlag.playlistsRebranding.enabled {
             changeNavTint(titleColor: nil, iconsColor: AppTheme.colorForStyle(.primaryIcon03))
         } else {
+            setupCloseButton()
             changeNavTint(titleColor: nil, iconsColor: AppTheme.colorForStyle(.primaryIcon02))
         }
         title = L10n.filterChoosePodcasts
         navigationController?.navigationBar.prefersLargeTitles = true
+        if FeatureFlag.playlistsRebranding.enabled {
+            navigationItem.largeTitleDisplayMode = .always
+        }
 
         let appearance = UINavigationBarAppearance()
         appearance.backgroundColor = AppTheme.colorForStyle(.primaryUi01)
@@ -187,7 +190,11 @@ class PodcastFilterOverlayController: PodcastChooserViewController, PodcastSelec
         filterToEdit.syncStatus = SyncStatus.notSynced.rawValue
         DataManager.sharedManager.save(filter: filterToEdit)
         NotificationCenter.postOnMainThread(notification: Constants.Notifications.filterChanged, object: filterToEdit)
-        dismiss(animated: true, completion: nil)
+        if FeatureFlag.playlistsRebranding.enabled {
+            navigationController?.popViewController(animated: true)
+        } else {
+            dismiss(animated: true, completion: nil)
+        }
 
         if !filterToEdit.isNew {
             Analytics.track(.filterUpdated, properties: ["group": "podcasts", "source": analyticsSource])
@@ -270,6 +277,7 @@ class PodcastFilterOverlayController: PodcastChooserViewController, PodcastSelec
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         if FeatureFlag.playlistsRebranding.enabled, indexPath.section == 0 {
             let cell = podcastTable.dequeueReusableCell(withIdentifier: podcastsSmartRuleHeaderCellId)!
+            cell.backgroundColor = AppTheme.colorForStyle(.primaryUi01)
             cell.contentView.backgroundColor = AppTheme.colorForStyle(.primaryUi01)
             cell.contentConfiguration = UIHostingConfiguration {
                 SmartRuleToggleHeaderView(viewModel: viewModel)
